+++ requirements.txt
@@ -1,3 +1,3 @@
-python-dotenv==1.0.0
-python-telegram-bot[job_queue]==20.5
+python-dotenv==1.0.1
+python-telegram-bot[job-queue]>=20.5
 requests==2.31.0
