+++ src/fantaformazionireminder/constants/notifications_interval.py
