+++ src/fantaformazionireminder/resources/seriea_calendar.csv
@@ -209,16 +209,16 @@ Match Number,Round Number,Date,Location,Home Team,Away Team,Result
 204,21,22/02/2024 20:45,Olimpico Grande Torino,Torino,Lazio,
 203,21,28/02/2024 18:00,Mapei Stadium - Città del Tricolore,Sassuolo,Napoli,
 209,21,28/02/2024 20:45,San Siro,Inter,Atalanta,
-212,22,26/01/2024 20:45,Unipol Domus,Cagliari,Torino,
-211,22,27/01/2024 15:00,Atleti Azzurri d'Italia,Atalanta,Udinese,
-215,22,27/01/2024 18:00,Juventus Stadium,Juventus,Empoli,
-217,22,27/01/2024 20:45,San Siro,Milan,Bologna,
-214,22,28/01/2024 12:30,Luigi Ferraris,Genoa,Lecce,
-218,22,28/01/2024 15:00,Brianteo,Monza,Sassuolo,
-220,22,28/01/2024 15:00,Marcantonio Bentegodi,Hellas Verona,Frosinone,
-216,22,28/01/2024 18:00,Olimpico,Lazio,Napoli,
-213,22,28/01/2024 20:45,Artemio Franchi,Fiorentina,Inter,
-219,22,29/01/2024 20:45,Arechi,Salernitana,Roma,
+212,22,26/01/2024 20:45,Unipol Domus,Cagliari,Torino,1 - 2
+211,22,27/01/2024 15:00,Atleti Azzurri d'Italia,Atalanta,Udinese,2 - 0
+215,22,27/01/2024 18:00,Juventus Stadium,Juventus,Empoli,1 - 1
+217,22,27/01/2024 20:45,San Siro,Milan,Bologna,2 - 2
+214,22,28/01/2024 12:30,Luigi Ferraris,Genoa,Lecce,2 - 1
+218,22,28/01/2024 15:00,Brianteo,Monza,Sassuolo,1 - 0
+220,22,28/01/2024 15:00,Marcantonio Bentegodi,Hellas Verona,Frosinone,1 - 1
+216,22,28/01/2024 18:00,Olimpico,Lazio,Napoli,0 - 0
+213,22,28/01/2024 20:45,Artemio Franchi,Fiorentina,Inter,0 - 1
+219,22,29/01/2024 20:45,Arechi,Salernitana,Roma,1 - 2
 226,23,02/02/2024 20:45,Ettore Giardiniero - Via del mare,Lecce,Fiorentina,
 223,23,03/02/2024 15:00,Carlo Castellani,Empoli,Genoa,
 230,23,03/02/2024 15:00,Friuli,Udinese,Monza,
