+++ src/fantaformazionireminder/utils/expiry_message.py
